Add: a holoportOS release expression
This is tested to work on OSX (darwin) inside docker. This can form a basis for
our release. Note that I had to comment out ```system_tarball_pc``` due to some
errors, and if this is what we need to ship to the manufacturer, then we'll have
to uncomment it and troubleshoot.